requirements.txt
python-dotenv==1.0.1
python-telegram-bot==13.0
requests==2.32.3
